Add PythonService executable
