+++ README.md
@@ -1,2 +1,8 @@
 # ProgressDevOpsTask
 ProgressDevOpsTask
+
+
+- Install python from Windows Marketplace
+- Disable windows defender for your own good
+- set execution policy to "Unrestricted"
+- run pip install pyinstaller
