+++ install-uninstall.ps1
@@ -1,29 +1,33 @@
 $serviceName = "PythonService"
-$scriptPath = "C:\Users\valushka\service.py"
-$pythonExe = "C:\Users\valushka\PythonServiceVenv\Scripts\python.exe"
+$pythonExe = ".\PythonService.exe"
 $regPath = "HKLM:\SOFTWARE\PythonService"
+$cwd = (Get-Item .).FullName
 
 function Create-RegistryEntries {
     if (!(Test-Path $regPath)) {
         New-Item -Path $regPath -Force | Out-Null
     }
-    Set-ItemProperty -Path $regPath -Name "LogFilePath" -Value "C:\Windows\Temp\service.log"
+    Set-ItemProperty -Path $regPath -Name "LogFilePath" -Value "${cwd}\PythonService.log"
     Set-ItemProperty -Path $regPath -Name "LogInterval" -Value 10
     Write-Host "Registry keys created successfully."
 }
 
 function Install-Service {
+    Create-RegistryEntries
+
     # Install the service using pywin32
-    & $pythonExe $scriptPath install
+    & $pythonExe install
     
     # Set service to run as LocalService
-    sc.exe config $serviceName obj= "NT AUTHORITY\LocalService"
+    sc.exe config $serviceName obj= "NT SERVICE\$serviceName"
+
+    # Grant permissions for all files in the working folder
+    icacls "$cwd" /grant "NT SERVICE\${serviceName}:(RX)"
 
     if ($LASTEXITCODE -eq 0) {
         Write-Host "Service installed successfully."
         Start-Service -Name $serviceName
         Write-Host "Service started."
-        Create-RegistryEntries
     } else {
         Write-Host "Failed to create service. Ensure you're running this script as Administrator."
     }
@@ -31,7 +35,7 @@ function Install-Service {
 
 function Uninstall-Service {
     Stop-Service -Name $serviceName -Force -ErrorAction SilentlyContinue
-    & $pythonExe $scriptPath remove
+    & $pythonExe remove
     Remove-Item -Path $regPath -Recurse -Force -ErrorAction SilentlyContinue
     Write-Host "Service uninstalled successfully."
 }
